# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Fitness Tracker API",
    "version": "1.0.0",
    "description": "API documentation for the Fitness Tracker application"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/auth/token": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Get JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": ["ADMIN", "WRITER", "VISITOR"]
                  },
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": ["READ", "WRITE", "DELETE", "UPDATE"]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/exercises": {
      "get": {
        "tags": ["Exercises"],
        "summary": "Get all exercises",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of exercises"
          }
        }
      },
      "post": {
        "tags": ["Exercises"],
        "summary": "Create new exercise",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "gifUrl": {
                    "type": "string"
                  },
                  "target": {
                    "type": "string"
                  },
                  "bodyPart": {
                    "type": "string"
                  },
                  "equipment": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Exercise created successfully"
          }
        }
      }
    },
    "/api/workouts": {
      "get": {
        "tags": ["Workouts"],
        "summary": "Get all workouts for the authenticated user",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of workouts"
          }
        }
      },
      "post": {
        "tags": ["Workouts"],
        "summary": "Create new workout",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "exercises": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "exercise": {
                          "type": "string"
                        },
                        "sets": {
                          "type": "integer"
                        },
                        "reps": {
                          "type": "integer"
                        },
                        "weight": {
                          "type": "number"
                        }
                      }
                    }
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workout created successfully"
          }
        }
      }
    }
  }
}